Write findplayer
Challenge Completed!
